fix: satisfy backend-size and pkg-file-count mergeability gates
Move allOf ref-property helpers into generated_contract_common_test.go instead of adding a 16th contracttests file, and fold factory schema parity tests into parity_inventory_test.go so both packages stay within the 15-file pkg-file-count limit.
